requirements.txt
faiss-cpu==1.8.0.post1
requests==2.32.3
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
numpy==1.26.4